requirements.txt
datasets
transformers
wandb
pyyaml
tqdm
numpy
Pillow
accelerate
python-dotenv
hf_transfer
scikit-learn
